tbox/tdf/tdt: pin OnFile fields to explicit-width types
Replace bare long/short with int32_t/int16_t in TextBoxOnFile (==36),
ThreeDFontF (==12), and ThreeDTextF (==24). Static_asserts verify.

Project 1 audit step 3.

+++ src/engine/tbox.cpp
@@ -1784,14 +1784,14 @@ void TextBoxGobject::AssertValid(ulong grf)
 //
 struct TextBoxOnFile
 {
-    short bo;
-    short osk;
-    long nfrmFirst;
-    long nfrmMax;
-    long xpLeft;
-    long xpRight;
-    long ypTop;
-    long ypBottom;
+    int16_t bo;
+    int16_t osk;
+    int32_t nfrmFirst;
+    int32_t nfrmMax;
+    int32_t xpLeft;
+    int32_t xpRight;
+    int32_t ypTop;
+    int32_t ypBottom;
     ChildChunkID chid;
     bool fStory;
 };
